tests: better assertion messages
Currently the tests would always fail because of the 'Hypervisor' key
being wrong. And moreover, the failure message was useless because it
was impossible to see exactly which piece of the huge pile of data was
wrong.

Ultimately this is actually because of lxml. Even though it's dealing in
XML (which is all Unicode) it "helpfully" converts ASCII strings down to
str in Python 2. This means that most of the keys and values in the
output of beaker-system-scan are str not unicode. However when we load
the expected values with json.load() they come back as unicode not str.
So when unittest tries to diff the two structures they appear completely
different, even if there is only one tiny value incorrect.

Change-Id: I87043254aebafc6826ea1f471b3cb772fbaa9861